Java Assignment
1.  Why Add Employee class. In real life example this is how it would be
done. Now although it could have been achieved using the Collection
framework to store info about the employee (like the Name, Hour Rate,
Hours worked) but for me it just doesn't separate the logic well and
doesn't define the Employee as a separate entity.
2. Why JUnit 4.10 not 4.12
because 4.12 has caused: initializationError did a small research about
it, tried few things and it didn’t work. So I chose not to put too much
time on it.
3. I separated getTotalPay from calcTotalPay just to separate the user
friendly output from the calculation of the totalpay. And also so that
we can easily unit test the calculation function.